@@ -4,13 +4,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#if NETCOREAPP_3_0_GREATER
+#if NET
 using System;
 using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
+using System.Runtime.Intrinsics.Arm;
 using System.Runtime.Intrinsics.X86;
 using System.Numerics;
-using System.Runtime.CompilerServices;
 using nuint_64 = System.UInt64;
 using nuint_32 = System.UInt32;
 
@@ -50,7 +51,7 @@ namespace DotNetty.Common.Internal
             // If we got here, it means we saw some non-ASCII data, so within our
             // vectorized code paths below we'll handle all non-surrogate UTF-16
             // code points branchlessly. We'll only branch if we see surrogates.
-            // 
+            //
             // We still optimistically assume the data is mostly ASCII. This means that the
             // number of UTF-8 code units and the number of scalars almost matches the number
             // of UTF-16 code units. As we go through the input and find non-ASCII
@@ -63,7 +64,11 @@ namespace DotNetty.Common.Internal
             long tempUtf8CodeUnitCountAdjustment = 0;
             int tempScalarCountAdjustment = 0;
 
-            if (Sse2.IsSupported)
+            // Per https://github.com/dotnet/runtime/issues/41699, temporarily disabling
+            // ARM64-intrinsicified code paths. ARM64 platforms may still use the vectorized
+            // non-intrinsicified 'else' block below.
+
+            if (/* (AdvSimd.Arm64.IsSupported && BitConverter.IsLittleEndian) || */ Sse2.IsSupported)
             {
                 if (inputLength >= Vector128<ushort>.Count)
                 {
@@ -72,53 +77,87 @@ namespace DotNetty.Common.Internal
                     Vector128<short> vector8800 = Vector128.Create(unchecked((short)0x8800));
                     Vector128<ushort> vectorZero = Vector128<ushort>.Zero;
 
+                    Vector128<byte> bitMask128 = BitConverter.IsLittleEndian ?
+                        Vector128.Create(0x80402010_08040201).AsByte() :
+                        Vector128.Create(0x01020408_10204080).AsByte();
+
                     do
                     {
-                        Vector128<ushort> utf16Data = Sse2.LoadVector128((ushort*)pInputBuffer); // unaligned
-                        uint mask;
-
-                        // The 'charIsNonAscii' vector we're about to build will have the 0x8000 or the 0x0080
-                        // bit set (but not both!) only if the corresponding input char is non-ASCII. Which of
-                        // the two bits is set doesn't matter, as will be explained in the diagram a few lines
-                        // below.
+                        Vector128<ushort> utf16Data;
+                        if (AdvSimd.Arm64.IsSupported)
+                        {
+                            utf16Data = AdvSimd.LoadVector128((ushort*)pInputBuffer); // unaligned
+                        }
+                        else
+                        {
+                            utf16Data = Sse2.LoadVector128((ushort*)pInputBuffer); // unaligned
+                        }
 
                         Vector128<ushort> charIsNonAscii;
-                        if (Sse41.IsSupported)
+
+                        if (AdvSimd.Arm64.IsSupported)
                         {
-                            // sets 0x0080 bit if corresponding char element is >= 0x0080
+                            // Sets the 0x0080 bit of each element in 'charIsNonAscii' if the corresponding
+                            // input was 0x0080 <= [value]. (i.e., [value] is non-ASCII.)
+                            charIsNonAscii = AdvSimd.Min(utf16Data, vector0080);
+                        }
+                        else if (Sse41.IsSupported)
+                        {
+                            // Sets the 0x0080 bit of each element in 'charIsNonAscii' if the corresponding
+                            // input was 0x0080 <= [value]. (i.e., [value] is non-ASCII.)
                             charIsNonAscii = Sse41.Min(utf16Data, vector0080);
                         }
                         else
                         {
-                            // sets 0x8000 bit if corresponding char element is >= 0x0080
-                            charIsNonAscii = Sse2.AndNot(vector0080, Sse2.Subtract(vectorZero, Sse2.ShiftRightLogical(utf16Data, 7)));
+                            // Sets the 0x0080 bit of each element in 'charIsNonAscii' if the corresponding
+                            // input was 0x0080 <= [value] <= 0x7FFF. The case where 0x8000 <= [value] will
+                            // be handled in a few lines.
+
+                            charIsNonAscii = Sse2.AndNot(Sse2.CompareGreaterThan(vector0080.AsInt16(), utf16Data.AsInt16()).AsUInt16(), vector0080);
                         }
 
 #if DEBUG
-                        // Quick check to ensure we didn't accidentally set both 0x8080 bits in any element.
-                        uint debugMask = (uint)Sse2.MoveMask(charIsNonAscii.AsByte());
-                        Debug.Assert((debugMask & (debugMask << 1)) == 0, "Two set bits shouldn't occur adjacent to each other in this mask.");
+                        // Quick check to ensure we didn't accidentally set the 0x8000 bit of any element.
+                        uint debugMask;
+                        if (AdvSimd.Arm64.IsSupported)
+                        {
+                            debugMask = GetNonAsciiBytes(charIsNonAscii.AsByte(), bitMask128);
+                        }
+                        else
+                        {
+                            debugMask = (uint)Sse2.MoveMask(charIsNonAscii.AsByte());
+                        }
+                        Debug.Assert((debugMask & 0b_1010_1010_1010_1010) == 0, "Shouldn't have set the 0x8000 bit of any element in 'charIsNonAscii'.");
 #endif // DEBUG
 
-                        // sets 0x8080 bits if corresponding char element is >= 0x0800
-                        Vector128<ushort> charIsThreeByteUtf8Encoded = Sse2.Subtract(vectorZero, Sse2.ShiftRightLogical(utf16Data, 11));
+                        // Sets the 0x8080 bits of each element in 'charIsNonAscii' if the corresponding
+                        // input was 0x0800 <= [value]. This also handles the missing range a few lines above.
 
-                        mask = (uint)Sse2.MoveMask(Sse2.Or(charIsNonAscii, charIsThreeByteUtf8Encoded).AsByte());
+                        Vector128<ushort> charIsThreeByteUtf8Encoded;
+                        uint mask;
 
-                        // Each odd bit of mask will be 1 only if the char was >= 0x0080,
-                        // and each even bit of mask will be 1 only if the char was >= 0x0800.
+                        if (AdvSimd.IsSupported)
+                        {
+                            charIsThreeByteUtf8Encoded = AdvSimd.Subtract(vectorZero, AdvSimd.ShiftRightLogical(utf16Data, 11));
+                            mask = GetNonAsciiBytes(AdvSimd.Or(charIsNonAscii, charIsThreeByteUtf8Encoded).AsByte(), bitMask128);
+                        }
+                        else
+                        {
+                            charIsThreeByteUtf8Encoded = Sse2.Subtract(vectorZero, Sse2.ShiftRightLogical(utf16Data, 11));
+                            mask = (uint)Sse2.MoveMask(Sse2.Or(charIsNonAscii, charIsThreeByteUtf8Encoded).AsByte());
+                        }
+
+                        // Each even bit of mask will be 1 only if the char was >= 0x0080,
+                        // and each odd bit of mask will be 1 only if the char was >= 0x0800.
                         //
                         // Example for UTF-16 input "[ 0123 ] [ 1234 ] ...":
                         //
-                        //            ,-- set if char[1] is non-ASCII
-                        //            |   ,-- set if char[0] is non-ASCII
+                        //            ,-- set if char[1] is >= 0x0800
+                        //            |   ,-- set if char[0] is >= 0x0800
                         //            v   v
-                        // mask = ... 1 1 1 0
-                        //              ^   ^-- set if char[0] is >= 0x0800
-                        //              `-- set if char[1] is >= 0x0800
-                        //
-                        // (If the SSE4.1 code path is taken above, the meaning of the odd and even
-                        // bits are swapped, but the logic below otherwise holds.)
+                        // mask = ... 1 1 0 1
+                        //              ^   ^-- set if char[0] is non-ASCII
+                        //              `-- set if char[1] is non-ASCII
                         //
                         // This means we can popcnt the number of set bits, and the result is the
                         // number of *additional* UTF-8 bytes that each UTF-16 code unit requires as
@@ -137,9 +176,16 @@ namespace DotNetty.Common.Internal
                         // Surrogates need to be special-cased for two reasons: (a) we need
                         // to account for the fact that we over-counted in the addition above;
                         // and (b) they require separate validation.
-
-                        utf16Data = Sse2.Add(utf16Data, vectorA800);
-                        mask = (uint)Sse2.MoveMask(Sse2.CompareLessThan(utf16Data.AsInt16(), vector8800).AsByte());
+                        if (AdvSimd.Arm64.IsSupported)
+                        {
+                            utf16Data = AdvSimd.Add(utf16Data, vectorA800);
+                            mask = GetNonAsciiBytes(AdvSimd.CompareLessThan(utf16Data.AsInt16(), vector8800).AsByte(), bitMask128);
+                        }
+                        else
+                        {
+                            utf16Data = Sse2.Add(utf16Data, vectorA800);
+                            mask = (uint)Sse2.MoveMask(Sse2.CompareLessThan(utf16Data.AsInt16(), vector8800).AsByte());
+                        }
 
                         if (mask != 0)
                         {
@@ -164,7 +210,15 @@ namespace DotNetty.Common.Internal
                             //   Since 'mask' already has 00 in these positions (since the corresponding char
                             //   wasn't a surrogate), "mask AND mask2 == 00" holds for these positions.
 
-                            uint mask2 = (uint)Sse2.MoveMask(Sse2.ShiftRightLogical(utf16Data, 3).AsByte());
+                            uint mask2;
+                            if (AdvSimd.Arm64.IsSupported)
+                            {
+                                mask2 = GetNonAsciiBytes(AdvSimd.ShiftRightLogical(utf16Data, 3).AsByte(), bitMask128);
+                            }
+                            else
+                            {
+                                mask2 = (uint)Sse2.MoveMask(Sse2.ShiftRightLogical(utf16Data, 3).AsByte());
+                            }
 
                             // 'lowSurrogatesMask' has its bits occur in pairs:
                             // - 01 if the corresponding char was a low surrogate char,
@@ -433,6 +487,21 @@ namespace DotNetty.Common.Internal
             utf8CodeUnitCountAdjustment = tempUtf8CodeUnitCountAdjustment;
             scalarCountAdjustment = tempScalarCountAdjustment;
             return pInputBuffer;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static uint GetNonAsciiBytes(Vector128<byte> value, Vector128<byte> bitMask128)
+        {
+            Debug.Assert(AdvSimd.Arm64.IsSupported);
+
+            Vector128<byte> mostSignificantBitIsSet = AdvSimd.ShiftRightArithmetic(value.AsSByte(), 7).AsByte();
+            Vector128<byte> extractedBits = AdvSimd.And(mostSignificantBitIsSet, bitMask128);
+
+            // self-pairwise add until all flags have moved to the first two bytes of the vector
+            extractedBits = AdvSimd.Arm64.AddPairwise(extractedBits, extractedBits);
+            extractedBits = AdvSimd.Arm64.AddPairwise(extractedBits, extractedBits);
+            extractedBits = AdvSimd.Arm64.AddPairwise(extractedBits, extractedBits);
+            return extractedBits.AsUInt16().ToScalar();
         }
     }
 }
